Adds avx2 implementation for qtesla (#553)
* Renamed qtesla portable folder, to better integrate avx2 files.

* Added avx2 support to qtesla

* Disabled KAT check for qtesla when avx2 is enabled.

* Commented the disabling of kat check for qtesla w/ avx2

* Fixed portable qtesla version and disabled avx2 on windows.

* Now reads use options from oqsconfig.h, accessible in CI tests.

* Skip config option check on windows.

* Moved cshake128x4 from qtesla dir to common sha3 dir.

* Added qtesla avx2 KATs, and use them when avx2 is enabled

* Prettyprint

* Fixed the qtesla avx2 condition check (needed a not xor)

+++ src/sig/qtesla/avx2/pack.c
@@ -110,7 +110,7 @@ static void decode_pk(int32_t *pk, unsigned char *seedA, const unsigned char *pk
 
 static void encode_sig(unsigned char *sm, unsigned char *c, poly z) { // Encode signature sm
 	unsigned int i, j = 0;
-	uint64_t *t = (uint64_t *) z;
+	uint32_t *t = (uint32_t *) z;
 	uint32_t *pt = (uint32_t *) sm;
 
 	for (i = 0; i < (PARAM_N * (PARAM_B_BITS + 1) / 32); i += 10) {
@@ -214,7 +214,7 @@ static void decode_pk(int32_t *pk, unsigned char *seedA, const unsigned char *pk
 
 static void encode_sig(unsigned char *sm, unsigned char *c, poly z) { // Encode signature sm
 	unsigned int i, j = 0;
-	uint64_t *t = (uint64_t *) z;
+	uint32_t *t = (uint32_t *) z;
 	uint32_t *pt = (uint32_t *) sm;
 
 	for (i = 0; i < (PARAM_N * (PARAM_B_BITS + 1) / 32); i += 11) {
